I am tired of these motherfucking hitboxes on this motherfucking platforms
The meaning of life pertains to the significance of living or existence in general, and is sought through the question "What is the meaning of life?" Many other related questions include: "Why are we here?", "What is life all about?", or "What is the purpose of existence?" There have been many proposed answers to these questions from many different cultural and ideological backgrounds. The search for life's meaning has produced much philosophical, scientific, theological, and metaphysical speculation throughout history. Different people and cultures believe different things for the answer to this question.